requirements.txt
ragas
datasets
python-dotenv
unstructured
unstructured[pdf]
